+++ __integration-tests__/jest-puppeteer.config.js
@@ -7,11 +7,14 @@ module.exports = {
       launchTimeout: 10000,
     },
   ],
-  launch: {
-    // headless: false,
-  },
+  // launch: {
+  //   headless: false,
+  // },
+
+  // To run on chrome
+  // /Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome --remote-debugging-port=9222 --no-first-run --no-default-browser-check --user-data-dir=$(mktemp -d -t 'chrome-remote_data_dir')
   // connect: {
   //   browserWSEndpoint:
-  //     'ws://127.0.0.1:9222/devtools/browser/363eec36-cb96-43a6-ae75-414542742899',
+  //     'ws://127.0.0.1:9222/devtools/browser/9ade75ab-631a-4504-a7ee-e9610fdbb606',
   // },
 };